requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
holidays==0.67
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.10
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
multitasking==0.0.11
narwhals==1.27.1
numpy==1.26.4
openai==1.63.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
propcache==0.3.0
prophet==1.1.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycoingecko==3.2.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stanio==0.5.1
statsmodels==0.14.4
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
yfinance==0.2.54
zstandard==0.23.0
